requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
redis
faststream[rabbit]
aio-pika
python-dotenv
httpx
pytest
